Create a simple logger for output (#25)
`Thrain` previously had written all output directly to `stdout` using
`printf`. This made it hard to silence or redirect output.

This creates a simple logger with four levels, and updates the program
to write with this instead `printf`. Output is written almost
identically to how `printf` works, including a format string with
variadic arguments.

A set of comprehensive tests ensures the logger works as expected.

It is in principle possible to silence all output or to redirect it to a
local file instead of the screen. A future update should add that
feature to the command line args.